Turn into class, add setup.py for pip
